backend/requirements.txt
requests>=2.31.0
python-dotenv>=1.0.1
Pillow>=10.3.0
psycopg2-binary>=2.9.9
qdrant-client>=1.7.0
